refactor(DAO-2135): move disclosure and shared components into subfolders
Move BtcVaultDisclosureSection, DisclosureContent, and
BtcVaultWalletDisconnectedSection into disclosure/ subfolder.
Move ReviewRow into shared/ subfolder. Add barrel index.ts files.
