+++ src/app/btc-vault/BtcVaultPage.tsx
@@ -7,9 +7,9 @@ import { SectionContainer } from '@/app/communities/components/SectionContainer'
 import { ActiveRequestSection } from './ActiveRequestSection'
 import { BtcVaultBanners } from './BtcVaultBanners'
 import { BtcVaultDashboard } from './components/BtcVaultDashboard'
-import { BtcVaultDisclosureSection } from './components/BtcVaultDisclosureSection'
+import { BtcVaultDisclosureSection } from './components/disclosure/BtcVaultDisclosureSection'
 import { BtcVaultMetrics } from './components/BtcVaultMetrics'
-import { BtcVaultWalletDisconnectedSection } from './components/BtcVaultWalletDisconnectedSection'
+import { BtcVaultWalletDisconnectedSection } from './components/disclosure/BtcVaultWalletDisconnectedSection'
 import { CapitalAllocationSection } from './components/capital-allocation/CapitalAllocationSection'
 import { useActiveRequests } from './hooks/useActiveRequests'
