+++ src/app/btc-vault/components/disclosure/BtcVaultDisclosureSection.tsx
