+++ src/app/btc-vault/components/disclosure/DisclosureContent.tsx
