+++ src/app/btc-vault/components/disclosure/index.ts
@@ -1,0 +1,3 @@
+export { BtcVaultDisclosureSection } from './BtcVaultDisclosureSection'
+export { BtcVaultWalletDisconnectedSection } from './BtcVaultWalletDisconnectedSection'
+export { DisclosureContent } from './DisclosureContent'
