+++ src/app/btc-vault/components/shared/index.ts
@@ -1,0 +1,1 @@
+export { ReviewRow, type ReviewRowProps } from './ReviewRow'
